Fix stable app crash due to missing nav components
